requirements.txt
streamlit==1.37.0
pandas==2.2.2
fpdf2==2.8.5